fix: input half/max button style change to use general component TokenInput.tsx
